+++ src/level_files_manager.rs
@@ -11,7 +11,7 @@ use crate::level_files_upgrading::upgrade_level_file;
 use crate::level_index::YoleckLevelIndexEntry;
 use crate::{
     YoleckEditorState, YoleckEntryHeader, YoleckKnobsCache, YoleckLevelIndex, YoleckLoadingCommand,
-    YoleckManaged, YoleckRawEntry, YoleckRawLevel, YoleckState, YoleckTypeHandlers,
+    YoleckManaged, YoleckRawEntry, YoleckRawLevel, YoleckState,
 };
 
 const EXTENSION: &str = ".yol";
@@ -39,7 +39,6 @@ pub fn level_files_manager_section(world: &mut World) -> impl FnMut(&mut World, 
         Commands,
         ResMut<YoleckState>,
         ResMut<YoleckEditorLevelsDirectoryPath>,
-        Res<YoleckTypeHandlers>,
         Query<(Entity, &YoleckManaged)>,
         Res<State<YoleckEditorState>>,
         ResMut<NextState<YoleckEditorState>>,
@@ -66,7 +65,6 @@ pub fn level_files_manager_section(world: &mut World) -> impl FnMut(&mut World, 
             mut commands,
             mut yoleck,
             mut levels_directory,
-            yoleck_type_handlers,
             yoleck_managed_query,
             editor_state,
             mut next_editor_state,
@@ -84,22 +82,16 @@ pub fn level_files_manager_section(world: &mut World) -> impl FnMut(&mut World, 
             YoleckRawLevel::new(app_format_version, {
                 yoleck_managed_query
                     .iter()
-                    .map(|(_entity, yoleck_managed)| {
-                        let _handler = yoleck_type_handlers
-                            .type_handlers
-                            .get(&yoleck_managed.type_name)
-                            .unwrap();
-                        YoleckRawEntry {
-                            header: YoleckEntryHeader {
-                                type_name: yoleck_managed.type_name.clone(),
-                                name: yoleck_managed.name.clone(),
-                            },
-                            data: yoleck_managed
-                                .components_data
-                                .iter()
-                                .map(|(k, v)| (k.to_string(), v.to_owned()))
-                                .collect(),
-                        }
+                    .map(|(_entity, yoleck_managed)| YoleckRawEntry {
+                        header: YoleckEntryHeader {
+                            type_name: yoleck_managed.type_name.clone(),
+                            name: yoleck_managed.name.clone(),
+                        },
+                        data: yoleck_managed
+                            .components_data
+                            .iter()
+                            .map(|(k, v)| (k.to_string(), v.to_owned()))
+                            .collect(),
                     })
             })
         };